Multiuser support

+++ app/Http/Controllers/SchedulledBookingsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 use App\Models\SchedulledBooking;
 use App\Models\KronoxSession;
@@ -16,12 +17,19 @@ class SchedulledBookingsController extends Controller
 
     public function index()
     {
-        $bookings = SchedulledBooking::all();
+        $bookings = SchedulledBooking::where(['user' => Auth::user()->username])->get();
         $sessions = KronoxSession::where('sessionActive', 1)->orderBy('MdhUsername', 'asc')->get();
         return view('schedulledbookings', compact(['bookings', 'sessions']));
     }
 
-    public function delete(SchedulledBooking $booking){
+    public function delete(SchedulledBooking $booking)
+    {
+        if($booking->user != Auth::user()->username)
+        {
+            flash('You did not make that schedulled booking...')->error();
+            return redirect(action('SchedulledBookingsController@index'));
+        }
+
         $booking->delete();
         flash('Deleted Schedulled Booking');
         return redirect(action('SchedulledBookingsController@index'));
